+++ src/components/generic/workflow_config_form/configForm.jsx
@@ -110,7 +110,7 @@ const ConfigForm = ({
             {...props}
           />
         );
-      case 'keyValue':
+      case 'mapping':
         return (
           <KeyValueMapper
             key={id}
